requirements.txt
base45==0.4.4
colorama==0.4.6
Pillow==10.4.0
Requests==2.32.3
